Add Stubbs the Zombie autosplitter
I had a .asl lying around for the old 2005 port since forever. A new 2021 port came out, so I've added (partial) support for that now too. And figured I should add it to the autosplitters list for convenience.

If you are adding or modifying an Auto Splitter, please fill out this checklist:
* [x]  My Auto Splitter does not contain any malicious functionality. (Any kind of abuse in an Auto Splitter will result in an immediate ban.)

* [x]  I am not replacing an existing Auto Splitter by a different author, or I have received permission from the author to replace the existing Auto Splitter.