Updating Pepsiman Auto-Splitter URL
"Pepsiman Auto-Splitter,asl" had a space in its name, which didn't look very nice when downloaded. Re-uploaded the file to my repo with an underscore instead of a space and changed the url here for my peace of mind.

Also the description had an ampersand that wasn't showing correctly, so I just changed it for a regular "and".