Add a new Path of Exile splitter
Add a new autosplitter for path of exile, one already exists but hasn't been updated in over a year, and is missing some routing. I will continue to update my script with new routes. For the time being it is following the most common speedrun routing for act 10 all labs and skill points.