Update MindSeize autosplitter URL
This patch updates the MindSeize autosplitter URL to point to the new location of the repository. Although the URL currently works as intended, there's no way of knowing if this will change in the future if my old username is claimed by a different user at some point. As proof that this is still the same repository, the redirect to the new URL can be seen by attempting to navigate to https://github.com/shockkolate/mindseize-autosplitter and seeing that it redirects to https://github.com/ori-sky/mindseize-autosplitter .