Autosplitter for Desperados III
This game brings you to the level selection screen upon completing a level, but levels should be completed in order in full game runs so the splitter has logic to account for this. Also, the game's level name string contains several components including difficulty setting, so the splitter has parsing to grab just the relevant piece.