Update LiveSplit.AutoSplitters.xml
The original author of this autosplitter is no longer active in the community and lost the source. Unfortunately the existing autosplitter isn't compatible with the latest version of the game that came out over a year ago.

This PR updates to point to a new version that does work with the latest version of Spelunky as well as being hosted in a repo adjacent to the source which how been restored from decompiling.