Assassins Creed Liberation Remastered Autosplitter (#1668)
* Assassins Creed Liberation Remastered Autosplitter

Fixed description for the syndicate one so it no longer includes the dlc title as it works for the dlc and base game for steam and Ubisoft connect. Also added a autosplitter for Assassins creed Liberation Remastered. But note on speedrun.com and in turn for live split you put put the game on Assassins Creed III Remastered and then select the category as Liberation any%/100%. So that's why i didn't put the game as Assassins Creed Liberation Remastered.

* Update LiveSplit.AutoSplitters.xml

---------

Co-authored-by: wooferzfg <spapushin@gmail.com>