Change Doom 2016 Repo
This change updates the existing autosplitter by replacing it with another repository per the request of the original author. The readme of the original repo reads: "These scripts are no longer maintained. If something needs to be updated, fork the repo or copy the script somewhere else, host it yourself, and point livesplit to the new version."
Original repo readme can be found at https://github.com/drtchops/asl/blob/master/README.md